项目结构优化、持久层编写
- 将JNI、Macro逻辑独立到core包，core包用于存放核心功能逻辑，
- service controller dao domain等包面向MVC逻辑
- client包面向原先基于JavaFX编写的内嵌式客户端
- 重构了宏配置、宏实例的配置文件结构，提供多宏实例支持
- 为宏配置、宏实例提供基于JSON文件读写的持久层DAO
- 移除了JNIUtils，其逻辑被合并进入PlatformFocusMonitor

+++ src/main/java/club/pineclone/gtavops/domain/dto/macro/PercentageDTO.java
@@ -1,6 +1,6 @@
 package club.pineclone.gtavops.domain.dto.macro;
 
-public record PercentageDTO(Double percentage, Double min, Double max) {
+public record PercentageDTO(Double percentage, Integer min, Integer max) {
 
     public double parseDouble() {
         return min + percentage * (max - min);
@@ -10,12 +10,12 @@ public record PercentageDTO(Double percentage, Double min, Double max) {
         return (long) Math.floor(parseDouble());
     }
 
-    public static PercentageDTO ofValue(double value, double min, double max) {
+    public static PercentageDTO ofValue(double value, int min, int max) {
         double pct = (value - min) / (max - min);
         return new PercentageDTO(pct, min, max);
     }
 
-    public static PercentageDTO fromRealValue(long value, double min, double max) {
+    public static PercentageDTO fromRealValue(long value, int min, int max) {
         return ofValue(value, min, max);
     }
