+++ src/main/java/club/pineclone/gtavops/domain/vo/macro/PercentageVO.java
@@ -1,4 +1,4 @@
 package club.pineclone.gtavops.domain.vo.macro;
 
-public record PercentageVO (Double percentage, Double min, Double max) {
+public record PercentageVO (Double percentage, Integer min, Integer max) {
 }
